Remove use of the Friends API
The Friends app and API is being removed from Ubuntu Touch as it wasn't
designed to work under confinement and has some other issues as well.
The only thing it was used for was for friendly relative time strings,
which can easily be done with qml-extras. For simplicity, this is
wrapped in a fake object which works like Friends.

- Replace the friends object with a QtObject that links to
DateUtils.friendlyTime